try robustness updates

+++ pkgs/nixos/profiles/personal.nix
@@ -59,28 +59,46 @@ with import ../dependencies.nix; {
   }) // {
     environment.systemPackages = [
       (pkgs.writeShellScriptBin "anix-init" ''
-        set -euo pipefail
-
         make-title -c yellow "Setting up rcrsync"
-        read -rp "Enter the char key to unlock the rclone config: " CFGKEY
-        mkdir -p $HOME/.config/rclone && cd $HOME/.config/rclone
-        cp ${anixpkgs.pkgData.records.rcloneConf.data} ${anixpkgs.pkgData.records.rcloneConf.name}
-        sunnyside -s 0 -k $CFGKEY -t ${anixpkgs.pkgData.records.rcloneConf.name}
-        rm ${anixpkgs.pkgData.records.rcloneConf.name}
+
+        DO_RCLONE=y
+        if [[ -f $HOME/.config/rclone/rclone.conf ]]; then
+          read -rp "rclone config already found, proceed anyway? (y|n): " DO_RCLONE
+        fi
+        if [[ "$DO_RCLONE" == "y" ]]; then
+          read -rp "Enter the char key to unlock the rclone config: " CFGKEY
+          rm -rf $HOME/.config/rclone
+          mkdir -p $HOME/.config/rclone && cd $HOME/.config/rclone
+          cp ${anixpkgs.pkgData.records.rcloneConf.data} ${anixpkgs.pkgData.records.rcloneConf.name}
+          sunnyside -s 0 -k $CFGKEY -t ${anixpkgs.pkgData.records.rcloneConf.name}
+          rm ${anixpkgs.pkgData.records.rcloneConf.name}
+        else
+          echo_yellow "Skipping rclone config step"
+        fi
+
         cd $HOME
-        rcrsync init configs
-        rcrsync init secrets
-        rcrsync init data
-        rcrsync init documents
-        rcrsync init games
-        rcrsync init games2
+        rcrsync -v init configs
+        rcrsync -v init secrets
+        rcrsync -v init data
+        rcrsync -v init documents
+        rcrsync -v init games
+        rcrsync -v init games2
 
         make-title -c yellow "Setting up SSH and Nix"
-        rm -rf $HOME/.ssh
-        cp -r $HOME/data/.ssh $HOME/.ssh
-        cd $HOME/.ssh
-        fix-perms .
-        cd ..
+        DO_SSH=y
+        if [[ -d $HOME/.ssh ]]; then
+          read -rp ".ssh directory already present, proceed anyway? (y|n): " DO_SSH
+        fi
+        if [[ "$DO_SSH" == "y" ]]; then
+          rm -rf $HOME/.ssh
+          cp -r $HOME/data/.ssh $HOME/.ssh
+          cd $HOME/.ssh
+          fix-perms .
+          cd ..
+        else
+          echo_yellow "Skipping SSH config setup"
+        fi
+        
         sudo nix-channel --add https://nixos.org/channels/nixos-${nixos-version} nixpkgs
         sudo nix-channel --add https://github.com/nix-community/home-manager/archive/release-${nixos-version}.tar.gz home-manager
         sudo nix-channel --update
